Update organisation schools page
