+++ app/views/hiring_staff/organisations/schools/index.html.haml
@@ -4,9 +4,9 @@
   .govuk-grid-column-two-thirds
 
     %h1.govuk-heading-xl{ class: 'govuk-!-margin-bottom-0' }
-      = t('.title')
+      = t('.title', organisation_type: organisation_type_basic(current_organisation))
     %p
-      = t('.description')
+      = t('.description', organisation_type: organisation_type_basic(current_organisation))
 
     = render 'school_group'
 
@@ -21,4 +21,4 @@
       %h3.govuk-heading-m
         = t('.missing_information_title')
       %p
-        = t('.missing_information_description_html', mail_to: mail_to(t('help.email'), t('.contact_support'), class: 'govuk-link'))
+        = t('.missing_information_description_html', organisation_type: organisation_type_basic(current_organisation), mail_to: mail_to(t('help.email'), t('.contact_support'), class: 'govuk-link'))
